Hash translation, package, csv warning
Changed translations to use hash to help with references to other mods.
This makes it easier to fix up other mods that reference cards in Mod
in this mod.  A good example is Actions, which have not Unique ID,
and some mods only provide Default text and no key.  The hash
makes it so that the mod can use a hash based Localization key without
needing to lookup the other mod's key.  It will be constant.

Added a CSV load error to help diagnose issues when there is not a
second column in the csv.

Fixed package to use release build